feat: rename transaction eval and add error enum
