+++ crates/uplc/src/tx.rs
@@ -4,20 +4,20 @@ use pallas_primitives::{
 };
 use pallas_traverse::{Era, MultiEraTx};
 
-use crate::Error;
+use error::Error;
+use script_context::{ResolvedInput, SlotConfig};
 
-use self::script_context::{ResolvedInput, SlotConfig};
-
+mod error;
 mod eval;
 pub mod script_context;
 mod to_plutus_data;
 
-pub fn eval_tx(
+pub fn eval(
     tx: &MintedTx,
     utxos: &[ResolvedInput],
     cost_mdls: Option<&CostMdls>,
     slot_config: &SlotConfig,
-) -> anyhow::Result<Vec<Redeemer>> {
+) -> Result<Vec<Redeemer>, Error> {
     let redeemers = tx.transaction_witness_set.redeemer.as_ref();
 
     let lookup_table = eval::get_script_and_datum_lookup_table(tx, utxos);
@@ -41,7 +41,7 @@ pub fn eval_tx(
     }
 }
 
-pub fn eval_tx_raw(
+pub fn eval_raw(
     tx_bytes: &[u8],
     utxos_bytes: &[(Vec<u8>, Vec<u8>)],
     cost_mdls_bytes: &[u8],
@@ -68,7 +68,7 @@ pub fn eval_tx_raw(
     };
 
     match multi_era_tx {
-        MultiEraTx::Babbage(tx) => match eval_tx(&tx, &utxos, Some(&cost_mdls), &sc) {
+        MultiEraTx::Babbage(tx) => match eval(&tx, &utxos, Some(&cost_mdls), &sc) {
             Ok(redeemers) => Ok(redeemers
                 .iter()
                 .map(|r| r.encode_fragment().unwrap())
@@ -96,7 +96,7 @@ mod tests {
     };
     use pallas_traverse::{Era, MultiEraTx};
 
-    use super::{eval_tx, ResolvedInput, SlotConfig};
+    use super::{eval, ResolvedInput, SlotConfig};
 
     #[test]
     fn test_eval() {
@@ -322,7 +322,7 @@ mod tests {
             .unwrap();
         match multi_era_tx {
             MultiEraTx::Babbage(tx) => {
-                let redeemers = eval_tx(&tx, &utxos, Some(&cost_mdl), &slot_config).unwrap();
+                let redeemers = eval(&tx, &utxos, Some(&cost_mdl), &slot_config).unwrap();
 
                 assert_eq!(redeemers.len(), 1)
             }
@@ -556,7 +556,7 @@ mod tests {
             .unwrap();
         match multi_era_tx {
             MultiEraTx::Babbage(tx) => {
-                let redeemers = eval_tx(&tx, &utxos, Some(&cost_mdl), &slot_config).unwrap();
+                let redeemers = eval(&tx, &utxos, Some(&cost_mdl), &slot_config).unwrap();
 
                 println!("{:?}", redeemers.len());
             }
@@ -626,7 +626,7 @@ mod tests {
             .unwrap();
         match multi_era_tx {
             MultiEraTx::Babbage(tx) => {
-                let redeemers = eval_tx(&tx, &utxos, Some(&cost_mdl), &slot_config).unwrap();
+                let redeemers = eval(&tx, &utxos, Some(&cost_mdl), &slot_config).unwrap();
 
                 println!("{:?}", redeemers.len());
             }
