+++ crates/uplc/src/tx/error.rs
@@ -1,0 +1,7 @@
+use crate::machine;
+
+#[derive(thiserror::Error, Debug)]
+pub enum Error {
+    #[error("Runtime error")]
+    Machine(#[from] machine::Error),
+}
